The tranform operators where templated with the complex variables. For a multitude of reasons they should be templated with the precision rather

+++ src/modules/CPUSimulator.cpp
@@ -309,7 +309,7 @@ CPUSimulator::Simulate () {
 		p = rf; q = p; r = p;
 		a = zeros<cxfl> (size(rf));
 
-		an = pow(creal(norm(p)), 2.0);
+		an = pow(creal(norm(p)), (float)2.0);
 		rn = an;
 		
 		for (iters = 0; iters < m_sb->cgit; iters++) { 	// CG loop
